separate test framework into it's own file in preparation for it becoming it's own module
